tests/builders: Split into dedicated files
